CXO-2951: Moving Beacon3100 message preparation to UPL style

+++ upl/com.energyict.mdc.upl.impl/protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/rtu3/beacon3100/messages/firmwareobjects/DeviceInfo.java
@@ -19,9 +19,9 @@ public class DeviceInfo {
     private TypedProperties generalProperties;
     private TypedProperties dialectProperties;
     private List<SecurityProperty> securityProperties;
-    private int deviceID;
+    private long deviceID;
 
-    public DeviceInfo(TypedProperties generalProperties, TypedProperties dialectProperties, List<SecurityProperty> securityProperties, int deviceID) {
+    public DeviceInfo(TypedProperties generalProperties, TypedProperties dialectProperties, List<SecurityProperty> securityProperties, long deviceID) {
         this.generalProperties = generalProperties;
         this.dialectProperties = dialectProperties;
         this.securityProperties = securityProperties;
@@ -33,7 +33,7 @@ public class DeviceInfo {
     }
 
     @XmlAttribute
-    public int getDeviceID() {
+    public long getDeviceID() {
         return deviceID;
     }
